[REFACTOR] Design 관련 Repository 경로 변경 (#44)
