+++ src/main/java/ject/componote/domain/design/dao/DesignRepository.java
@@ -1,5 +1,6 @@
-package ject.componote.domain.design.domain;
+package ject.componote.domain.design.dao;
 
+import ject.componote.domain.design.domain.Design;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface DesignRepository extends JpaRepository<Design, Long> {
