+++ src/main/java/ject/componote/domain/design/dao/DesignSystemRepository.java
@@ -1,5 +1,6 @@
-package ject.componote.domain.design.domain;
+package ject.componote.domain.design.dao;
 
+import ject.componote.domain.design.domain.Design;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface DesignSystemRepository extends JpaRepository<Design, Long> {
