+++ src/main/java/ject/componote/domain/design/dao/link/DesignLinkRepository.java
@@ -1,5 +1,6 @@
-package ject.componote.domain.design.domain.link;
+package ject.componote.domain.design.dao.link;
 
+import ject.componote.domain.design.domain.link.DesignLink;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
